made .env file, renamed base.js to firebase.js
