+++ client/src/components/Dashboard.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import app from "../base"
+import app from "../firebase"
 
 const Dashboard = () => {
     return (
